+++ saleor/static/dashboard-next/customers/views/CustomerDetails.tsx
@@ -18,6 +18,7 @@ import { TypedCustomerDetailsQuery } from "../queries";
 import { RemoveCustomer } from "../types/RemoveCustomer";
 import { UpdateCustomer } from "../types/UpdateCustomer";
 import {
+  customerAddressesUrl,
   customerListUrl,
   customerRemovePath,
   customerRemoveUrl,
@@ -104,7 +105,9 @@ export const CustomerDetailsView: React.StatelessComponent<
                                   updateCustomerOpts.data.customerUpdate.errors
                               )}
                               saveButtonBar={formTransitionState}
-                              onAddressManageClick={() => undefined} // TODO: add address management #3173
+                              onAddressManageClick={() =>
+                                navigate(customerAddressesUrl(id))
+                              }
                               onBack={() => navigate(customerListUrl)}
                               onRowClick={id => navigate(orderUrl(id))}
                               onSubmit={formData =>