+++ saleor/static/dashboard-next/icons/Pages.tsx
@@ -3,9 +3,25 @@ import * as React from "react";
 
 export const Pages = createSvgIcon(
   <>
-    <rect width="24" height="24" fillOpacity="0" />
-    <rect width="18" height="18" fillOpacity="0" transform="translate(3 3)" />
-    <path d="M19 3H5C3.9 3 3 3.9 3 5V19C3 20.1 3.9 21 5 21H19C20.1 21 21 20.1 21 19V12V5C21 3.9 20.1 3 19 3ZM13 5H19V11H16L17 7L13 8V5ZM5 5H11V8L7 7L8 11H5V5ZM11 19H5V13H8L7 17L11 16V19ZM19 19H13V16L17 17L16 13H19V19ZM14.63 14.63L12 13.72L9.37 14.63L10.28 12L9.37 9.37L12 10.28L14.63 9.37L13.72 12L14.63 14.63Z" />
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M22.3617 7H35.5267L42.0006 13.4739V36.5548H28.2148V34.7644H40.2103V14.2155L40.012 14.0172H34.8V8.80523L34.7851 8.79036H24.1521V12.3573H22.3617V7ZM6 11.4613H19.165L25.6389 17.9352V41.0161H24.7437H6V11.4613ZM18.6 13.4283V19.4172H23.8485V39.2257H7.79036V13.2517H18.4234L18.6 13.4283ZM22.2 23.0172H9.6V25.7172H22.2V23.0172ZM9.6 27.5172H22.2V30.2172H9.6V27.5172ZM37.5 16.7172H28.5V19.4172H37.5V16.7172ZM28.5 21.2172H37.5V31.1172H28.5V21.2172ZM22.2 32.0172H9.6V34.7172H22.2V32.0172Z"
+      fill="url(#paint0_linear)"
+    />
+    <defs>
+      <linearGradient
+        id="paint0_linear"
+        x1="4.94116"
+        y1="7"
+        x2="35.6523"
+        y2="46.7403"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="#13BEBB" />
+        <stop offset="1" stop-color="#3EE7CD" />
+      </linearGradient>
+    </defs>
   </>
 );
 Pages.displayName = "Pages";