+++ tests/conftest.py
@@ -20,6 +20,7 @@ from saleor.account.models import Address, User
 from saleor.checkout import utils
 from saleor.checkout.models import Cart
 from saleor.checkout.utils import add_variant_to_cart
+from saleor.core.utils.taxes import DEFAULT_TAX_RATE_NAME
 from saleor.dashboard.menu.utils import update_menu
 from saleor.discount import VoucherType
 from saleor.discount.models import Sale, Voucher, VoucherTranslation
@@ -222,6 +223,7 @@ def shipping_zone_without_countries(db):  # pylint: disable=W0613
         shipping_zone=shipping_zone)
     return shipping_zone
 
+
 @pytest.fixture
 def shipping_method(shipping_zone):
     return ShippingMethod.objects.create(
@@ -316,7 +318,8 @@ def permission_manage_orders():
 @pytest.fixture
 def product_type(color_attribute, size_attribute):
     product_type = ProductType.objects.create(
-        name='Default Type', has_variants=True, is_shipping_required=True)
+        name='Default Type', has_variants=True, is_shipping_required=True,
+        tax_rate=DEFAULT_TAX_RATE_NAME)
     product_type.product_attributes.add(color_attribute)
     product_type.variant_attributes.add(size_attribute)
     return product_type
@@ -337,7 +340,8 @@ def product(product_type, category):
 
     product = Product.objects.create(
         name='Test product', price=Money('10.00', 'USD'),
-        product_type=product_type, attributes=attributes, category=category)
+        product_type=product_type, attributes=attributes, category=category,
+        tax_rate=DEFAULT_TAX_RATE_NAME)
 
     variant_attr = product_type.variant_attributes.first()
     variant_attr_value = variant_attr.values.first()